backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-multipart
pydantic
python-dotenv
jose
transformers
torch
sentence-transformers
pdf2image
pillow
opencv-python
pdfplumber
pdf2image
pillow
transformers
torch
sentence-transformers